Migrate cost tracking and estimation into common/cost
