+++ common/cost/BUILD.bazel
@@ -9,6 +9,8 @@ go_library(
     name = "go_default_library",
     srcs = [
         "cost.go",
+        "estimator.go",
+        "tracker.go",
     ],
     importpath = "cel.dev/cel-go/common/cost",
 )
@@ -18,6 +20,8 @@ go_test(
     size = "small",
     srcs = [
         "cost_test.go",
+        "estimator_test.go",
+        "tracker_test.go",
     ],
     embed = [
         ":go_default_library",
